# Add a "safe IO" flag