## Summary

Adds a flag to indicate all IO is thread-safe.  This doesn't help with programs that have a mix of thread-safe and thread-unsafe IO, but I hope that there is a reasonably large class of programs where the only IO is for things like thread-local PRNGs and the like.

**Related issues:** closes #224

## Checklist

- [x] Add new tests to dejafu-tests
- [x] Performance comparison on systematic testing for par monad example

## Benchmark results

### Without systematic par test

**Before:**

```
$ stack exec -- dejafu-tests +RTS -s

 142,901,146,320 bytes allocated in the heap
  31,388,207,960 bytes copied during GC
      14,024,608 bytes maximum residency (6323 sample(s))
         146,744 bytes maximum slop
              38 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     131731 colls,     0 par   70.915s  70.897s     0.0005s    0.0354s
  Gen  1      6323 colls,     0 par    4.964s   4.963s     0.0008s    0.0022s

  TASKS: 227991 (227986 bound, 5 peak workers (5 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time  171.291s  (163.647s elapsed)
  GC      time   75.879s  ( 75.860s elapsed)
  EXIT    time    0.001s  (  0.003s elapsed)
  Total   time  247.171s  (239.510s elapsed)

  Alloc rate    834,260,065 bytes per MUT second

  Productivity  69.3% of total user, 68.3% of total elapsed
```

**After:**

```
$ stack exec -- dejafu-tests +RTS -s

 143,989,514,888 bytes allocated in the heap
  31,970,755,160 bytes copied during GC
      14,022,152 bytes maximum residency (6201 sample(s))
         145,816 bytes maximum slop
              38 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     132926 colls,     0 par   72.848s  72.780s     0.0005s    0.0330s
  Gen  1      6201 colls,     0 par    4.891s   4.888s     0.0008s    0.0024s

  TASKS: 227859 (227854 bound, 5 peak workers (5 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time  169.741s  (162.130s elapsed)
  GC      time   77.738s  ( 77.668s elapsed)
  EXIT    time    0.001s  (  0.003s elapsed)
  Total   time  247.481s  (239.800s elapsed)

  Alloc rate    848,289,519 bytes per MUT second

  Productivity  68.6% of total user, 67.6% of total elapsed
```
### With systematic par test

**Before:**

```
$ stack exec -- dejafu-tests +RTS -s

 621,082,673,912 bytes allocated in the heap                                                                                          
 319,844,942,272 bytes copied during GC
     174,860,488 bytes maximum residency (8406 sample(s))
       1,697,592 bytes maximum slop
             489 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     586924 colls,     0 par   544.721s  544.390s     0.0009s    0.2977s
  Gen  1      8406 colls,     0 par    6.108s   6.104s     0.0007s    0.0019s

  TASKS: 241379 (241374 bound, 5 peak workers (5 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time  416.744s  (408.224s elapsed)
  GC      time  550.829s  (550.494s elapsed)
  EXIT    time    0.034s  (  0.042s elapsed)
  Total   time  967.607s  (958.760s elapsed)

  Alloc rate    1,490,321,894 bytes per MUT second

  Productivity  43.1% of total user, 42.6% of total elapsed
```

`ParMonad/testing exposes a deadlock/systematically` took 717.68s.

**After:**

```
$ stack exec -- dejafu-tests +RTS -s

 149,739,906,080 bytes allocated in the heap                                                                                          
  36,054,433,992 bytes copied during GC
      14,078,072 bytes maximum residency (6604 sample(s))
         197,448 bytes maximum slop
              38 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     138048 colls,     0 par   78.591s  78.554s     0.0006s    0.0376s
  Gen  1      6604 colls,     0 par    5.159s   5.161s     0.0008s    0.0029s

  TASKS: 228163 (228158 bound, 5 peak workers (5 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time  172.307s  (164.302s elapsed)
  GC      time   83.751s  ( 83.715s elapsed)
  EXIT    time    0.001s  (  0.003s elapsed)
  Total   time  256.059s  (248.020s elapsed)

  Alloc rate    869,028,162 bytes per MUT second

  Productivity  67.3% of total user, 66.2% of total elapsed
```

`ParMonad/testing exposes a deadlock/systematically` took 11.32s.